+++ ElementsKernel/src/Lib/Real.cpp
@@ -44,13 +44,13 @@ bool almostEqual2sComplement(const float& left, const float& right, const int& m
     int32_t a_int = *reinterpret_cast<const int32_t *>(&left);
     // Make a_int lexicographically ordered as a twos-complement int
     if (a_int < 0) {
-        a_int = 0x80000000 - a_int;
+        a_int = static_cast<int32_t>(0x80000000 - a_int);
     }
     // Make b_int lexicographically ordered as a twos-complement int
     //    int b_int = *(int*)&b;
     int32_t b_int = *reinterpret_cast<const int32_t *>(&right);
     if (b_int < 0) {
-        b_int = 0x80000000 - b_int;
+        b_int = static_cast<int32_t>(0x80000000 - b_int);
     }
     int32_t int_diff = abs(a_int - b_int);
     if (int_diff <= max_ulps && -max_ulps <= int_diff) {
@@ -69,13 +69,13 @@ bool almostEqual2sComplement(const double& left, const double& right, const int&
     int64_t a_int = *reinterpret_cast<const int64_t *>(&left);
     // Make a_int lexicographically ordered as a twos-complement int
     if (a_int < 0) {
-        a_int = 0x8000000000000000LL - a_int;
+        a_int = static_cast<int64_t>(0x8000000000000000LL - a_int);
     }
     // Make b_int lexicographically ordered as a twos-complement int
     //    long long b_int = *(long long*)&b;
     int64_t b_int = *reinterpret_cast<const int64_t *>(&right);
     if (b_int < 0) {
-        b_int = 0x8000000000000000LL - b_int;
+        b_int = static_cast<int64_t>(0x8000000000000000LL - b_int);
     }
     int64_t int_diff = abs(a_int - b_int);
     if (int_diff <= max_ulps && -max_ulps <= int_diff) {
